+++ src/main/java/org/example/hatealcohol/user/config/CorsMvcConfig.java
@@ -1,4 +1,4 @@
-package org.example.hatealcohol.config;
+package org.example.hatealcohol.user.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
