+++ src/main/java/org/example/hatealcohol/user/dto/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package org.example.hatealcohol.dto;
+package org.example.hatealcohol.user.dto;
 
 import java.util.ArrayList;
 import java.util.Collection;
