+++ src/main/java/org/example/hatealcohol/user/dto/OAuth2Response.java
@@ -1,4 +1,4 @@
-package org.example.hatealcohol.dto;
+package org.example.hatealcohol.user.dto;
 
 public interface OAuth2Response {
