+++ src/main/java/org/example/hatealcohol/user/dto/UserDTO.java
@@ -1,4 +1,4 @@
-package org.example.hatealcohol.dto;
+package org.example.hatealcohol.user.dto;
 
 import lombok.Getter;
 import lombok.Setter;
