+++ src/main/java/org/example/hatealcohol/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package org.example.hatealcohol.repository;
+package org.example.hatealcohol.user.repository;
 
-import org.example.hatealcohol.entity.User;
+import org.example.hatealcohol.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
